processes: prevent spamming UI with console windows
Projects such as Renode use CCompilerTask and CLinkerTask to build using
gcc in C#.  On Windows, this causes a new terminal window to be spawned
for every process, resulting in a storm of windows.

Set `CreateNoWindow=true` in order to prevent this storm of console
windows when building or cleaning using an IDE.

https://blogs.msdn.microsoft.com/jmstall/2006/09/28/how-to-start-a-console-app-in-a-new-window-the-parents-window-or-no-window/
has more information.

Signed-off-by: Sean Cross <sean@xobs.io>